WRCC Monitor Names and Unit IDs

The WRCC https://wrcc.dri.edu/cgi-bin/smoke.plFire Cache Smoke Monitor Archive provides access to a variety of monitors that can be accessed with the wrcc_createMonitorObject function. Use of this funciton requires a valid unitID. The WRCC object is a list of lists. The element named unitIDs is itself a list of three named vectors, each containing the unitIDs and associated names for one of the categories of monitors available at WRCC:

cache

miscellaneous

usfs_regional

Format

A list of lists

Details

WRCC monitor names and unitIDs

Note

This list of monitor IDs relfects unitIDs found on the WRCC site on June 12, 2019.