Add Elevation Data to a Dataframe

Google APIs are used to determine elevation associated with the locations specified by the lonVar and latVar columns of the incoming dataframe.

Address information is obtained by using the ggmap package.

addGoogleElevation(df, lonVar = "longitude", latVar = "latitude",
  existingMeta = NULL)

Arguments

df
dataframe with geolocation information (e.g. those created by wrcc_qualityControl() or airsis_qualityControl)

lonVar
name of longitude variable in the incoming dataframe

latVar
name of the latitude variable in the incoming dataframe

existingMeta
existing 'meta' dataframe from which to obtain metadata for known monitor deployments

Value

Input dataframe with (possibly) additional column: elevation.

References

https://developers.google.com/maps/documentation/elevation/intro