Create AIRSIS Site Location Metadata Dataframe

After an AIRSIS tibble has been enhanced with additional columns generated by addClustering we are ready to pull out site information associated with unique deployments.

These will be rearranged into a dataframe organized as deployment-by-property with one row for each monitor deployment.

This site information found in tbl is augmented so that we end up with a uniform set of properties associated with each monitor deployment. The list of columns in the returned meta dataframe is:

> names(p$meta)
 [1] "monitorID"             "longitude"             "latitude"
 [4] "elevation"             "timezone"              "countryCode"
 [7] "stateCode"             "siteName"              "agencyName"
[10] "countyName"            "msaName"               "monitorType"
[13] "monitorInstrument"     "aqsID"                 "pwfslID"
[16] "pwfslDataIngestSource" "telemetryAggregator"   "telemetryUnitID"

airsis_createMetaDataframe(tbl, provider = as.character(NA),
  unitID = as.character(NA), pwfslDataIngestSource = "AIRSIS",
  existingMeta = NULL, addGoogleMeta = FALSE, addEsriMeta = FALSE)

Arguments

tbl
single site AIRSIS tibble after metadata enhancement

provider
identifier used to modify baseURL ['APCD'|'USFS']

unitID
character or numeric AIRSIS unit identifier

pwfslDataIngestSource
identifier for the source of monitoring data, e.g. 'AIRSIS'

existingMeta
existing 'meta' dataframe from which to obtain metadata for known monitor deployments

addGoogleMeta
logicial specifying wheter to use Google elevation and reverse geocoding services

addEsriMeta
logicial specifying wheter to use ESRI elevation and reverse geocoding services

Value

A meta dataframe for use in a ws_monitor object.

See also