Identify AIRSIS monitor type

Examine the column names of the incoming dataframe (or first line of raw text) to identify different types of monitor data provided by AIRSIS.

The return is a list includes everything needed to identify and parse the raw data using readr::read_csv():

monitorType -- identification string

rawNames -- column names from the data (including special characters)

columnNames -- assigned column names (special characters repaced with '.')

columnTypes -- column type string for use with readr::read_csv()

The monitorType will be one of:

"BAM1020" -- BAM1020 (e.g. USFS #49 in 2010)

"EBAM" -- EBAM (e.g. USFS #1026 in 2010)

"ESAM" -- E-Sampler (e.g. USFS #1002 in 2010)

"UNKOWN" -- ???

airsis_identifyMonitorType(df)

Arguments

df
dataframe or raw character string containing AIRSIS data

Value

List including monitorType, rawNames, columnNames and columnTypes.

References

Interagency Real Time Smoke Monitoring

Examples

if (FALSE) {
fileString <- airsis_downloadData( 20150701, 20151231, provider='USFS', unitID='1026')
monitorTypeList <- airsis_identifyMonitorType(fileString)
}