Create EPA data dataframe

After additional columns(i.e. datetime, and monitorID) have been applied to an EPA dataframe, we are ready to extract the PM2.5 values and store them in a data dataframe organized as time-by-monitor.

The first column of the returned dataframe is named datetime and contains a POSIXct time in UTC. Additional columns contain data for each separate monitorID.

epa_createDataDataframe(tbl)

Arguments

tbl
an EPA raw tibble after metadata enhancement

Value

A data dataframe for use in a ws_monitor object.