Create a dataframe of current monitor data

Extracts current status data from a ws_monitor object. In addition to monitor metadata, the returned data include the following:

monitorID - the PWFSL monitor ID

lastValidTime - UTC POSIXct time corresponding to the last valid pm25 datum

lastValidLocalTimestamp - ASCII version of lastValidTime in monitor-local timezone

processingTime - UTC POSIXct when the function is run (i.e. 'now')

datetime - UTC POSIXct of the datetime that the data represents. See datetime parameter. In most instances, the same as processingTime

latency - (difference between processingTime -1 hr) and lastValidTime, floored to the hour

lastValid_PM2.5_nowcast - NowCast value at lastValidTime

lastValid_PM2.5_1hr - PM2.5 value at lastValidTime (should never be null)

lastValid_PM2.5_3hr - mean of the three hours preceeding lastValidTime

yesterday_PM2.5_24hr - local time midnight-to-midnight 24-hour mean for the day prior to processingTime

monitoringSiteUrl - URL to PWFSL Monitoring v4 site, initialized with a specific monitor

The three-hour average will remove missing values and may represent an average of 1-3 hours.

The yesterday average will remove up to six missing values, returning NA if more than six hours of yesterday's data are missing.

Any monitors with no data from before the desired datetime will be excluded.

NOTE: yesterday_PM2.5_24hr represents yesterday relative to datetime -- i.e. actually yesterday. The various 'lastValid' values describe the most recent valid values -- before datetime, whenever they occurred.

monitor_currentData(ws_monitor, datetime = lubridate::now("UTC"),
  monitoringUrlBase = "http://tools.airfire.org/monitoring/v4/#!/?monitors=")

Arguments

ws_monitor
ws_monitor object

datetime
Time to which data will be 'current' (integer or character representing YYYYMMDDHH or POSIXct. If not POSIXct, interpreted as UTC time). So if datetime is 3 hours ago, a dataframe with the most current data from 3 hours ago will be returned.

monitoringUrlBase
base URL for constructing a link to the PWFSL Smoke Monitoring site

Value

A tibble of "lastValid" data and associated timing information.

Details

Data are assigned to the beginning of the hour they represent. So a PM2.5 value assigned to 2pm will represent data averaged over the period 14:00 - 14:59. This is in keeping with a day representing 00:00 - 23:59 and a month representing the beginning of the month until the end of the month.

Because of this, data for 2pm should never be available until just after 3pm. If it is currently 3:15pm then we need to subtract 1 hour from the current processingTime before subtracting the lastValidTime to generate the latency.

The parameter datetime is meant to simulate current data at a time different from the current time. Therefore, when datetime is specified, the returned dataframe will represent data up to an hour before datetime.

Examples

# NOT RUN {
wa <- loadLatest() %>% monitor_subset(stateCodes = 'WA')
wa_current <- monitor_currentData(wa)
# }