Load annual PM2.5 monitoring data

Wrapper function to load and combine annual data from AirNow, AIRSIS and WRCC.

If dataDir is defined, data will be loaded from this local dirctory. Otherwise, data will be loaded from the monitoring data repository maintained by PWFSL.

The annual files loaded by this function are updated on the 15'th of each month and cover the period from the beginning of the year to the end of the last month.

For data during the last 45 days, use monitor_loadDaily().

For the most recent data, use monitor_loadLatest().

Currently supported parameters include the following:

PM2.5

Avaialble RData files can be seen at: https://haze.airfire.org/monitoring/

monitor_loadAnnual(year = NULL, parameter = "PM2.5",
  baseUrl = "https://haze.airfire.org/monitoring", dataDir = NULL,
  aqsPreference = "airnow")

Arguments

year
Desired year (integer or character representing YYYY).

parameter
Parameter of interest.

baseUrl
Base URL for data files.

dataDir
Local directory containing 'daily' data files.

aqsPreference
Preferred data source for AQS data when annual data files are available from both `epa` and `airnow`.

Value

A ws_monitor object with PM2.5 monitoring data.

See also

Examples

# NOT RUN {
monitor_loadAnnual(2014) %>%
  monitor_subset(stateCodes='MT', tlim=c(20140801,20140901)) %>%
  monitor_map()
# }