Replace ws_monitor Data with Another Value

Use an R expression to identify values for replacement.

The RR expression given in filter is used to identify elements in ws_monitor$data that should be replaced. Typical usage would include

replacing negative values with 0

replacing unreasonably high values with NA

Expressions should use data for the left hand side of the comparison.

monitor_replaceData(ws_monitor, filter, value)

Arguments

ws_monitor
ws_monitor object

filter
an RR expression used to identify values for replacement

value
replacement value

Examples

wa <- monitor_subset(Northwest_Megafires, stateCodes='WA')
wa_zero <- monitor_replaceData(wa, data < 0, 0)