Calculate Time Averages

This function extracts the data dataframe from ws_monitor object and renames the 'datetime' column so that it can be processed by the openair package's timeAverage() function. (See that function for details.)

monitor_timeAverage(ws_monitor, ...)

Arguments

ws_monitor
ws_monitor object

...
additional arguments to be passed to openair::timeAverage()

Value

A ws_monitor object with data that have been proccessed by openair::timeAverage().

Examples

C_V <- monitor_subset(Carmel_Valley, tlim=c(2016080800,2016081023),
                      timezone='America/Los_Angeles')
C_V_3hourly <- monitor_timeAverage(C_V, avg.time="3 hour")
head(C_V$data, n=15)

#>               datetime 060530002_01
#> 1  2016-08-08 07:00:00            8
#> 2  2016-08-08 08:00:00           24
#> 3  2016-08-08 09:00:00           46
#> 4  2016-08-08 10:00:00           86
#> 5  2016-08-08 11:00:00          132
#> 6  2016-08-08 12:00:00          197
#> 7  2016-08-08 13:00:00          197
#> 8  2016-08-08 14:00:00          252
#> 9  2016-08-08 15:00:00          250
#> 10 2016-08-08 16:00:00           77
#> 11 2016-08-08 17:00:00           17
#> 12 2016-08-08 18:00:00            9
#> 13 2016-08-08 19:00:00            6
#> 14 2016-08-08 20:00:00            6
#> 15 2016-08-08 21:00:00            7

head(C_V_3hourly$data, n=5)

#>              datetime 060530002_01
#> 1 2016-08-08 07:00:00    26.000000
#> 2 2016-08-08 10:00:00   138.333333
#> 3 2016-08-08 13:00:00   233.000000
#> 4 2016-08-08 16:00:00    34.333333
#> 5 2016-08-08 19:00:00     6.333333