Write Monitor Data as CSV File

Writes the contents of the ws_monitor object as CSV file. By default, the output is a "human readable" CSV that includes both meta and data. This format is useful for point-and-click spreadsheet users who want to have everything on a single sheet.

To obtain machine parseable CSV files you can use metaOnly or dataOnly which are mutually exclusive but which generate CSV files that can be automatically ingested.

monitor_writeCSV(ws_monitor, saveFile = "ws_monitor.csv",
  metaOnly = FALSE, dataOnly = FALSE)

Arguments

ws_monitor
ws_monitor object

saveFile
filename where CSV will be written

metaOnly
flag specifying whether to use ws_monitor$meta for a machine parseable CSV

dataOnly
flag specifying whether to use ws_monitor$data for a machine parseable CSV

Examples

# NOT RUN {
data("Carmel_Valley")
Carmel_Valley <- monitor_subset(Carmel_Valley, tlim=c(20160802,20160803))
monitor_writeCSV(Carmel_Valley, 'CarmelValley.csv')
monitor_writeCSV(Carmel_Valley, 'CarmelValley_meta.csv', metaOnly = TRUE)
monitor_writeCSV(Carmel_Valley, 'CarmelValley_data.csv', dataOnly = TRUE)
# }