Apply Quality Control to Raw WRCC E-Sampler Tibble

Perform various QC measures on WRCC EBAM data.

The any numeric values matching the following are converted to NA

x < -900

x == -9.9899

x == 99999

The following columns of data are tested against valid ranges:

Flow

AT

RHi

ConcHr

A POSIXct datetime column (UTC) is also added based on DateTime.

wrcc_ESAMQualityControl(tbl, valid_Longitude = c(-180, 180),
  valid_Latitude = c(-90, 90), remove_Lon_zero = TRUE,
  remove_Lat_zero = TRUE, valid_Flow = c(1.999, 2.001),
  valid_AT = c(-Inf, 150), valid_RHi = c(-Inf, 55),
  valid_Conc = c(-Inf, 5000), flagAndKeep = FALSE)

Arguments

tbl
single site tibble created by wrcc_parseData()

valid_Longitude
range of valid Longitude values

valid_Latitude
range of valid Latitude values

remove_Lon_zero
flag to remove rows where Longitude == 0

remove_Lat_zero
flag to remove rows where Latitude == 0

valid_Flow
range of valid Flow values

valid_AT
range of valid AT values

valid_RHi
range of valid RHi values

valid_Conc
range of valid ConcHr values

flagAndKeep
flag, rather than remove, bad data during the QC process

Value

Cleaned up tibble of WRCC monitor data.

See also