Obain AirNow data and create ws_monitor objects

This function uses the airnow_downloadParseData function to download monthly dataframes of AirNow data and restructures that data into a format that is compatible with the PWFSLSmoke package ws_monitor data model.

AirNow data parameters include at least the following list:

BARPR

BC

CO

NO

NO2

NO2Y

NO2X

NOX

NOOY

OC

OZONE

PM10

PM2.5

PRECIP

RHUM

SO2

SRAD

TEMP

UV-AETH

WD

WS

Setting parameters=NULL will generate a separate ws_monitor object for each of the above parameters.

airnow_createMonitorObjects(parameters = NULL,
  startdate = strftime(lubridate::now(), "%Y%m%d", tz = "UTC"),
  hours = 24, zeroMinimum = TRUE, addGoogleMeta = TRUE)

Arguments

parameters
vector of names of desired pollutants or NULL for all pollutants

startdate
desired start date (integer or character representing YYYYMMDD[HH])

hours
desired number of hours of data to assemble

zeroMinimum
logical specifying whether to convert negative values to zero

addGoogleMeta
logicial specifying wheter to use Google elevation and reverse geocoding services

Value

List where each element contains a ws_monitor object for a unique parameter (e.g: "PM2.5", "NOX").

Note

As of 2017-12-17, it appears that hourly data are available only for 2016 and not for earlier years.

See also

Examples

# NOT RUN {
monList <- airnow_createMonitorObjects(c("O3","PM2.5"), 20160701)
pm25 <- monList$PM2.5
o3 <- monList$O3
# }