Combine List of ws_monitor Objects into Single ws_monitor Object

Combines a list of one or more ws_monitor objects into a single ws_monitor object by merging the meta and data dataframes from each object in monitorList.

When monitorList contains only two ws_monitor objects the monitor_combine() function can be used to extend time ranges for monitorIDs that are found in both ws_monitor objects. This can be used to 'grow' a ws_monitor object by appending subsequent months or years. (Note, however, that this can be CPU intensive process.)

monitor_combine(monitorList)

Arguments

monitorList
list containing one or more ws_monitor objects

Value

A ws_monitor object combining all monitoring data from monitorList.

Examples

# NOT RUN {
monitorList <- list()
monitorList[[1]] <- airsis_createMonitorObject(20160701, 20161231, 'USFS', '1031')
monitorList[[2]] <- airsis_createMonitorObject(20160701, 20161231, 'USFS', '1032')
monitorList[[3]] <- airsis_createMonitorObject(20160701, 20161231, 'USFS', '1033')
monitorList[[4]] <- airsis_createMonitorObject(20160701, 20161231, 'USFS', '1034')
ws_monitor <- monitor_combine(monitorList)
monitor_leaflet(ws_monitor)
# }