Isolate Individual Monitors

Filters ws_monitor according to the parameters passed in. If any parameter is not specified, that parameter will not be used in the filtering.

After filtering, each monitorID found in ws_monitor is extracted and its data dataframe is restricted to the times from when that monitor first datapoint until its last datapoint.

This function is useful when ws_monitor objects are created for mobile monitors that are deployed to different locations in different years.

monitor_isolate(ws_monitor, xlim = NULL, ylim = NULL, tlim = NULL,
  monitorIDs = NULL, stateCodes = NULL, timezone = "UTC")

Arguments

ws_monitor
ws_monitor object

xlim
optional vector with low and high longitude limits

ylim
optional vector with low and high latitude limits

tlim
optional vector with start and end times (integer or character representing YYYYMMDD[HH] or POSIXct)

monitorIDs
optional vector of monitorIDs

stateCodes
optional vector of stateCodes

timezone
Olson timezone passed to link{parseDatetime} when parsing numeric tlim

Value

A list of isolated ws_monitor objects.

See also

Examples

N_M <- Northwest_Megafires
# monitor_leaflet(N_M) # to identify Spokane monitorIDs
Spokane <- monitor_subsetBy(N_M, stringr::str_detect(N_M$meta$monitorID,'^53063'))
Spokane$meta$monitorID

#> [1] "530630047_01" "530630021_01" "530639999_01" "530639997_01" "530639996_01"

monitorList <- monitor_isolate(Spokane)
names(monitorList)

#> [1] "530630047_01" "530630021_01" "530639999_01" "530639997_01" "530639996_01"