Create Timeseries Plot from a Raw Dataframe

Creates a plot of raw monitoring data as generated using raw_enhance().

Other options for parameter include "temperature", "humidity", "windSpeed", "windDir", "pressure" or any of the other raw parameters (try names(df) to see list of options)

rawPlot_timeseries(
  df,
  parameter = "pm25",
  tlim = NULL,
  localTime = TRUE,
  shadedNight = TRUE,
  shadedBackground = NULL,
  sbLwd = 1,
  add = FALSE,
  gridPos = "",
  gridCol = "black",
  gridLwd = 1,
  gridLty = "solid",
  dayLwd = 0,
  hourLwd = 0,
  hourInterval = 6,
  ...
)

Arguments

df
enhanced, raw dataframe as created by the raw_enhance() function

parameter
raw parameter to plot

tlim
optional vector with start and end times (integer or character representing YYYYMMDD[HH])

localTime
logical specifying whether tlim is in local time or UTC

shadedNight
add nighttime shading

shadedBackground
add vertical lines for a second parameter

sbLwd
shaded background line width

add
logical specifying whether to add to the current plot

gridPos
position of grid lines either 'over', 'under' or '' for no grid lines

gridCol
grid line color

gridLwd
grid line width

gridLty
grid line type

dayLwd
day marker line width

hourLwd
hour marker line width

hourInterval
interval for grid (max=12)

...
additional arguments to pass to lines() function

Details

Note that for multiple deployments, shadedNight defaults to use the lat/lon for the first deployment, which in theory could be somewhat unrepresentative, such as if deployments have a large range in latitude.

Note

If more than one timezone is found, localTime is ignored and UTC is used.