Identify WRCC monitor type

Examine the column names of the incoming character vector to identify different types of monitor data provided by WRCC.

The return is a list includes everything needed to identify and parse the raw data using readr::read_tsv():

monitorType -- identification string

rawNames -- column names from the data (including special characters)

columnNames -- assigned column names (special characters repaced with '.')

columnTypes -- column type string for use with readr::read_csv()

The monitorType will be one of:

"WRCC_TYPE1" -- ???

"WRCC_TYPE2" -- ???

"UNKOWN" -- ???

wrcc_identifyMonitorType(fileString)

Arguments

fileString
character string containing WRCC data

Value

List including monitorType, rawNames, columnNames and columnTypes.

References

WRCC Fire Cache Smoke Monitor Archive

Examples

# NOT RUN {
fileString <- wrcc_downloadData(20160701, 20160930, unitID='1307')
monitorTypeList <- wrcc_identifyMonitorType(fileString)
# }